+++ core/src/org/amityregion5/ZombieGame/client/screen/LoadingScreen.java
@@ -16,9 +16,7 @@ import com.badlogic.gdx.utils.Align;
  */
 
 /**
- *
  * @author sergeys
- *
  */
 public class LoadingScreen implements Screen {
 
@@ -34,8 +32,8 @@ public class LoadingScreen implements Screen {
 
 		batch.begin();
 
-		calibri30.draw(batch, "Loading" + (dots > 0 ? "." : "")
-				+ (dots > 1 ? "." : "") + (dots > 2 ? "." : ""), 10, 400, 1180, Align.center, false);
+		calibri30.draw(batch, "Loading" + (dots > 0 ? "." : "") + (dots > 1 ? "." : "") + (dots > 2 ? "." : ""), 10,
+				400, 1180, Align.center, false);
 
 		if (dCount > 0.5) {
 			dots++;
@@ -52,47 +50,42 @@ public class LoadingScreen implements Screen {
 	}
 
 	@Override
-	public void resize(int width, int height) {
-	}
+	public void resize(int width, int height) {}
 
 	@Override
 	public void show() {
 		FreeTypeFontGenerator generator = ZombieGame.instance.fontGenerator;
 		boolean disposeGenerator = false;
-		
+
 		if (generator == null) {
-			generator = new FreeTypeFontGenerator(
-					Gdx.files.internal("font/Helvetica.ttf"));
+			generator = new FreeTypeFontGenerator(Gdx.files.internal("font/Helvetica.ttf"));
 			disposeGenerator = true;
 		} else {
 			generator = ZombieGame.instance.fontGenerator;
 		}
 
 		FreeTypeFontParameter parameter = new FreeTypeFontParameter();
-		parameter.size = (int) (26);
+		parameter.size = (26);
 
 		calibri30 = generator.generateFont(parameter);
 
 		calibri30.setColor(1, 1, 1, 1);
 
 		batch = new SpriteBatch();
-		
+
 		if (disposeGenerator) {
 			generator.dispose();
 		}
 	}
 
 	@Override
-	public void hide() {
-	}
+	public void hide() {}
 
 	@Override
-	public void pause() {
-	}
+	public void pause() {}
 
 	@Override
-	public void resume() {
-	}
+	public void resume() {}
 
 	@Override
 	public void dispose() {
